Incorporation of an issues template
An issue template is incorporated, to request information from people who want to create a bug report.
The template asks for the data to be completed in order to efficiently create the report and be able to solve it, having the necessary information to be able to solve the problem.